requirements.txt
Flask
python-dotenv
azure-ai-documentintelligence
azure-storage-blob
scikit-learn
openai
gunicorn